projects/argos/lambda/requirements.txt
onnxruntime==1.16.3
numpy==1.26.4
pillow==11.3.0
ultralytics>=8.0

opencv-python-headless>=4.10.0.82

scipy>=1.11
redis>=5.0
boto3>=1.34
matplotlib>=3.8
